requirements.txt
aiofiles==24.1.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.10
isodate==0.6.1
lxml==5.3.0
numpy==2.1.1
outcome==1.3.0.post0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
PySocks==1.7.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==6.10.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0

